requirements.txt
aws-cdk-lib==2.147.0
constructs>=10.0.0,<11.0.0
aws-cdk.lambda-layer-kubectl-v30>=2.0.1
